Prevent opening or using chest before expected restock has executed
This commit attempts to fix a timing issue whereby players can duplicate
items using the chests.

Upon trying to open a restock chest, a check is implemented to see
whether the next restock has happened at the expected time or not (e.g.
if it has not, the time until next restock will be <0s, as it has not
updated). If not, the player is prevented from opening the chest.

When a chest is already open and a restock has or should have occurred,
the user is prevented from taking any items. This makes use of a
comparison between the times that the chest was last restocked, and also
a similar check to that of when the chest is opened, i.e. checks if the
time until next restock is negative.